Disable Mark FitNesse root folder action, since it is not store in a module file anyway.
